move speaker team member social media links form into separate component
