+++ src/pages/speakerTeamMemberSocialMediaLinks.svelte
@@ -4,105 +4,31 @@
 
     import type { LoadDashboard, LoadUserSocials } from 'types/dashboardLoadTypes';
     import type { DashboardSocialMediaLink } from 'types/dashboardProvideTypes';
-    import type {
-        SetAllUpdateSocialMediaLink, SetCreateSocialMediaLink,
-    } from 'types/dashboardSetTypes';
+    import type { SetAllUpdateSocialMediaLink, SetCreateSocialMediaLink } from 'types/dashboardSetTypes';
+    import type { DeleteReturn, ValidateReturn } from 'pageHelper/speakerTeamMemberSocials';
 
     import Tabs from 'elements/navigation/tabs.svelte';
     import UnsavedChangesCallbackWrapper from 'elements/navigation/unsavedChangesCallbackWrapper.svelte';
     import Message from 'elements/text/message.svelte';
     import SectionDashboard from 'elements/section/sectionDashboard.svelte';
-    import EditSocialMedia from 'elements/input/editSocialMedia.svelte';
-    import Button from 'elements/input/button.svelte';
-    import GeneralPopup from 'elements/popups/generalPopup.svelte';
     import SaveMessage from 'elements/text/saveMessage.svelte';
+    import SpeakerTeamMemberSocialMediaLinkForm from './speakerTeamMemberSocialMediaLinkForm.svelte';
 
-    import { error } from '@sveltejs/kit';
     import { trySaveDashboardDataAsync } from 'helper/trySaveDashboardData';
     import { isSuccessType, SaveMessageType } from 'types/saveMessageType';
-    import { setUnsavedChanges } from 'stores/saved';
-    import { apiUrl } from 'helper/links';
-    import { z } from 'zod';
-    import { loadSocials } from 'pageHelper/speakerTeamMemberSocials';
+    import { deleteLinkAsync, loadSocials } from 'pageHelper/speakerTeamMemberSocials';
+    import { getIDFromSocialMediaType, validate } from 'pageHelper/speakerTeamMemberSocials';
 
     export let data: LoadDashboard & LoadUserSocials;
 
-    let deletePopup: GeneralPopup;
     let saveMessage: SaveMessage;
     let messages: string[] = [];
 
-    function getIDFromSocialMediaType(type: string): number {
-        for (let element of data.socialTypes) {
-            if (type === element.name) {
-                return element.id;
-            }
-        }
-        console.error(`not able to look up requested social media link type ID: ${type}`);
-        throw error(500);
-    }
-
-    function validate(): string[] {
-        const https: string      = 'https://';
-        const http: string       = 'http://';
-        const messages: string[] = [];
-        for (let entry of data.socials) {
-            entry.url = entry.url.trim();
-            if (entry.url.length === 0) {
-                messages.push(`Ein ${entry.name} - Eintrag ist leer`);
-                continue;
-            }
-
-            if (!entry.url.startsWith(https) && !entry.url.startsWith(http)) {
-                entry.url = https + entry.url;
-            }
-
-            const validated = z.string().url().safeParse(entry.url);
-            if (!validated.success) {
-                messages.push(`${entry.url} ist keine valide URL.`);
-                continue; // keep this continue here to prevent bugs when one adds code after this.
-            }
-        }
-
-        data.socials = data.socials; // reassign for reactivity
-        return messages;
-    }
-
-    async function deleteLinkAsync(index: number): Promise<void> {
-        const id = data.socials[index].id;
-        if (id !== 0) {
-            const deleteResponse: Response = await fetch(
-                apiUrl(`/api/dashboard/user/social-media-link/${id}`),
-                { method: 'DELETE' },
-            );
-            if (!deleteResponse.ok) {
-                saveMessage.setSaveMessage(SaveMessageType.DeleteError);
-                console.error(`Deleting Link: Bad Backend response: ${deleteResponse.status}`);
-                return;
-            }
-        }
-
-        data.socials = data.socials.filter((item) => item.id !== id);
-        saveMessage.setSaveMessage(SaveMessageType.Delete);
-    }
-
-    function addLink(): void {
-        const newLink: DashboardSocialMediaLink = {
-            approved:             false,
-            id:                   0,
-            name:                 'Web',
-            url:                  '',
-            user_id:              data.roles.user_id,
-            requested_changes:    null,
-            social_media_type_id: getIDFromSocialMediaType('Web'),
-        };
-        data.socials                            = [
-            ...data.socials,
-            newLink,
-        ];
-    }
 
     async function trySaveAsync(): Promise<boolean> {
-        messages = validate();
+        const validateResult: ValidateReturn = validate(data.socials);
+        messages                             = validateResult.messages;
+        data.socials                         = validateResult.data;
         if (messages.length > 0) {
             return false;
         }
@@ -119,7 +45,10 @@
 
             toSave.social_media_links.push({
                                                id:                   link.id,
-                                               social_media_type_id: getIDFromSocialMediaType(link.name),
+                                               social_media_type_id: getIDFromSocialMediaType(
+                                                   data.socialTypes,
+                                                   link.name,
+                                               ),
                                                url:                  link.url,
                                            });
         }
@@ -145,7 +74,7 @@
 
     async function tryCreateAsync(link: DashboardSocialMediaLink): Promise<boolean> {
         const toSave: SetCreateSocialMediaLink = {
-            social_media_type_id: getIDFromSocialMediaType(link.name),
+            social_media_type_id: getIDFromSocialMediaType(data.socialTypes, link.name),
             url:                  link.url,
         };
 
@@ -157,6 +86,13 @@
 
         return isSuccessType(messageType);
     }
+
+    async function tryDeleteAsync(e: CustomEvent<number>): Promise<void> {
+        const value: number              = e.detail;
+        const deleteResult: DeleteReturn = await deleteLinkAsync(data.socials, value);
+        saveMessage.setSaveMessage(deleteResult.message);
+        data.socials = deleteResult.data;
+    }
 </script>
 
 <Tabs
@@ -165,22 +101,6 @@
       classes="navigation-tabs-dashboard-subpage"
 />
 <UnsavedChangesCallbackWrapper callback={trySaveAsync} />
-<GeneralPopup
-      bind:this={deletePopup}
-      headline="Link löschen?"
-      text="Gelöschte Links können nicht wiederhergestellt werden."
-      acceptButtonText="Löschen"
-      denyButtonText="Abbrechen"
-      acceptCallback={(value) => {
-		if (typeof value === 'number') {
-			deleteLinkAsync(value);
-			return;
-		}
-
-		console.error('provided entry index from generic popup is not number');
-	}}
-      denyCallback={() => {}}
-></GeneralPopup>
 
 <SectionDashboard classes="standard-dashboard-section">
     {#if data.roles.is_speaker && data.roles.is_team_member}
@@ -195,36 +115,9 @@
                  {message} />
     {/each}
 
-    <form on:submit|preventDefault={trySaveAsync}>
-        <EditSocialMedia
-              links={data.socials}
-              socialMediaTypes={data.socialTypes.map((x) => x.name)}
-              deleteCallback={(e) => {
-				deletePopup.show(e);
-			}}
-              on:input={setUnsavedChanges}
-        />
-        <div class="dashboard-social-media-links-button-wrapper">
-            <Button ariaLabel="Klicke hier, um einen neuen Link hinzuzufügen"
-                    on:click={addLink}
-            >Neu
-            </Button
-            >
-            <Button ariaLabel="Klicke hier, um die Eingaben zu speichern"
-                    type={'submit'}
-            >Speichern
-            </Button
-            >
-        </div>
-    </form>
-</SectionDashboard>
+    <SpeakerTeamMemberSocialMediaLinkForm bind:data={data}
+                                          on:save={trySaveAsync}
+                                          on:delete={tryDeleteAsync} />
 
-<style>
-    .dashboard-social-media-links-button-wrapper {
-        display:         flex;
-        flex-direction:  row;
-        gap:             var(--full-gap);
-        margin-top:      var(--4x-gap);
-        justify-content: center;
-    }
-</style>
+
+</SectionDashboard>
